+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 # Version 2.x
 
+# 2.0.0 (unreleased)
+
+* Removed `PropertyTypeArray`, which is superseeded by `PropertyTypeIterable`.
+* Removed the deprecated `PropertyTypeIterable::getCollectionClass`. Use `PropertyTypeIterable::getTraversableClass`
+* Removed the deprecated `PropertyTypeIterable::isCollection`. Use `PropertyTypeIterable::isTraversable`
+* `JMSTypeParser::getTraversableClass` returns `Traversable::class` instead of `Doctrine\Common\Collections\Collection` for general traversable properties.
+
+
 # Version 1.x
 
 # 1.2.0